requirements.txt
API telegram == 4.18.1
psycopg2-binary == 2.9.9
sqlalchemy == 2.0.30
pillow == 10.3.0